# openapi.yaml
openapi: 3.1.1
info:
  title: Open-Spot API
  description: |
    개인 장소 기록 및 관리 서비스 API

    ## 주요 기능
    - Google OAuth2 인증 (Bearer Token + HttpOnly Cookie)
    - 개인 장소 기록 관리 (평점, 리뷰, 태그)
    - 장소 그룹 관리
    - 푸시 알림 관리

    ## 인증 방식
    - **Access Token**: Authorization 헤더로 전달 (Bearer 방식)
    - **Refresh Token**: HttpOnly Cookie로 전달 (XSS 방어)

    ## MSA 아키텍처
    - **Gateway (8080)**: 모든 API의 진입점, JWT 검증 수행
    - **Auth Service (8081)**: Google OAuth2 로그인, JWT 발급
    - **Location Service (8082)**: 장소 관리
    - **Notification Service (8083)**: 알림 관리

    **중요**: 프론트엔드는 반드시 Gateway(8080)를 통해 API를 호출해야 합니다.
  version: 1.0.0
  contact:
    name: Open-Spot Team
servers:
  - url: http://localhost:8080
    description: Gateway (모든 API 호출은 이 URL 사용)
  - url: http://localhost:8081
    description: Auth Service (직접 호출 금지, Gateway 경유 필수)
  - url: http://localhost:8082
    description: Location Service (직접 호출 금지, Gateway 경유 필수)
  - url: http://localhost:8083
    description: Notification Service (직접 호출 금지, Gateway 경유 필수)
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/notifications/tokens:
    post:
      tags:
        - Notifications
      summary: 디바이스 토큰 등록
      description: FCM 푸시 알림을 위한 디바이스 토큰을 등록합니다.
      operationId: registerDeviceToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterTokenRequest'
      responses:
        '200':
          description: 토큰 등록 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RegisterTokenResponse'
  /api/v1/notifications:
    get:
      tags:
        - Notifications
      summary: 알림 목록 조회
      description: 사용자의 알림 목록을 페이지네이션으로 조회합니다.
      operationId: getNotifications
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 알림 목록 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/NotificationPageResponse'
...
  /api/v1/notifications/{notificationId}/read:
    put:
      tags:
        - Notifications
      summary: 알림 읽음 처리
      description: 특정 알림을 읽음 상태로 변경합니다.
      operationId: markNotificationAsRead
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 알림 읽음 처리 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/notifications/settings:
    get:
      tags:
        - Notifications
      summary: 알림 설정 조회
      description: 사용자의 알림 설정을 조회합니다.
      operationId: getNotificationSettings
      responses:
        '200':
          description: 알림 설정 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/NotificationSettingsResponse'

    put:
      tags:
        - Notifications
      summary: 알림 설정 업데이트
      description: 사용자의 알림 설정을 업데이트합니다.
      operationId: updateNotificationSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSettingsRequest'
      responses:
        '200':
          description: 알림 설정 업데이트 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/NotificationSettingsResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Access Token을 Authorization 헤더로 전달합니다.

        **사용법**: `Authorization: Bearer <access-token>`

        **토큰 획득**:
        1. Google 로그인 성공 시 URL 쿼리 파라미터로 전달
        2. 토큰 갱신 API 응답 본문에서 획득

        **인증 플로우 (MSA 표준)**:
        1. 클라이언트 → Gateway: `Authorization: Bearer <token>` 헤더 포함
        2. Gateway: JWT 검증 (HS256 알고리즘)
        3. Gateway: 검증 성공 시 `X-User-Id` 헤더 추가
        4. 내부 서비스: `X-User-Id` 헤더로 사용자 식별 (JWT 검증 불필요)
  parameters:
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
...
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        data:
          type: object
          nullable: true
        error:
          type: object
          nullable: true
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    PageInfo:
      type: object
      required:
        - number
        - size
        - totalElements
        - totalPages
        - first
        - last
      properties:
        number:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean

    # ========================================
    # Authentication Schemas
    # ========================================
...
    RegisterTokenRequest:
      type: object
      required:
        - token
        - deviceType
      properties:
        token:
          type: string
        deviceType:
          type: string
          enum: [WEB, ANDROID, IOS]
        deviceId:
          type: string
          nullable: true
    RegisterTokenResponse:
      type: object
      required:
        - id
        - token
        - deviceType
        - isActive
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        token:
          type: string
        deviceType:
          type: string
          enum: [WEB, ANDROID, IOS]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    NotificationResponse:
      type: object
      required:
        - id
        - title
        - body
        - notificationType
        - isRead
        - isSent
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
        notificationType:
          type: string
          enum: [REPORT_COMPLETE, SYSTEM_NOTICE]
        referenceId:
          type: string
          format: uuid
          nullable: true
        isRead:
          type: boolean
        isSent:
          type: boolean
        createdAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
          nullable: true
    NotificationPageResponse:
      type: object
      required:
        - content
        - page
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/NotificationResponse'
        page:
          $ref: '#/components/schemas/PageInfo'
...
    NotificationSettingsResponse:
      type: object
      required:
        - reportEnabled
        - systemEnabled
      properties:
        reportEnabled:
          type: boolean
        systemEnabled:
          type: boolean
    UpdateSettingsRequest:
      type: object
      required:
        - reportEnabled
        - systemEnabled
      properties:
        reportEnabled:
          type: boolean
        systemEnabled:
          type: boolean